begin move mpn_mod to its own module
