+++ doc/source/index.rst
@@ -139,6 +139,7 @@ Integers mod n
    nmod_poly_factor.rst
    nmod_mpoly.rst
    nmod_mpoly_factor.rst
+   mpn_mod.rst
    fmpz_mod.rst
    fmpz_mod_vec.rst
    fmpz_mod_mat.rst
