+++ doc/source/index_integers_mod.rst
@@ -16,6 +16,7 @@
        nmod_poly_factor.rst
        nmod_mpoly.rst
        nmod_mpoly_factor.rst
+       mpn_mod.rst
        fmpz_mod.rst
        fmpz_mod_vec.rst
        fmpz_mod_mat.rst
