+++ src/gr/mpn_mod.c
@@ -18,6 +18,8 @@
 #include "gr_vec.h"
 #include "gr_mat.h"
 
+#include "mpn_mod.h"
+
 /* for wide add_ssss.... macros. todo; these ought to be provided
    everywhere */
 #if FLINT_BITS == 64 && defined(__AVX2__)
@@ -28,133 +30,6 @@
 /* todo: fast polynomial multiplication */
 /* todo: powering */
 /* todo: special functions */
-
-
-
-#define UI_ABS_SI(x) (((slong)(x) < 0) ? (-(ulong)(x)) : ((ulong)(x)))
-
-/* Single limbs are already dealt with well by nmod, and excluding them
-   allows avoiding various special cases. */
-#define MPN_MOD_MIN_LIMBS 2
-
-/* This should be small enough that we can stack-allocate mpn_mods
-   and temporary buffers a small multiple of the size.
-
-   For bigger sizes, use fmpz_mod.
-
-   Before resizing this, make sure that any lookup tables that go up
-   to MPN_MOD_MAX_LIMBS (such as tuning tables) are large enough. */
-#define MPN_MOD_MAX_LIMBS 16
-
-typedef struct
-{
-    mp_size_t nlimbs;
-    mp_limb_t d[MPN_MOD_MAX_LIMBS];
-    mp_limb_t dinv[MPN_MOD_MAX_LIMBS];
-    mp_limb_t dnormed[MPN_MOD_MAX_LIMBS];
-    flint_bitcnt_t norm;
-    truth_t is_prime;
-}
-_mpn_mod_ctx_struct;
-
-#define MPN_MOD_CTX(ctx) ((_mpn_mod_ctx_struct *)(GR_CTX_DATA_AS_PTR(ctx)))
-#define MPN_MOD_CTX_NLIMBS(ctx) (MPN_MOD_CTX(ctx)->nlimbs)
-#define MPN_MOD_CTX_MODULUS(ctx) (MPN_MOD_CTX(ctx)->d)
-#define MPN_MOD_CTX_MODULUS_NORMED(ctx) (MPN_MOD_CTX(ctx)->dnormed)
-#define MPN_MOD_CTX_MODULUS_PREINV(ctx) (MPN_MOD_CTX(ctx)->dinv)
-#define MPN_MOD_CTX_NORM(ctx) (MPN_MOD_CTX(ctx)->norm)
-#define MPN_MOD_CTX_IS_PRIME(ctx) (MPN_MOD_CTX(ctx)->is_prime)
-
-FLINT_FORCE_INLINE
-int flint_mpn_equal_p(mp_srcptr x, mp_srcptr y, mp_size_t xsize)
-{
-    slong i;
-    for (i = 0; i < xsize; i++)
-    {
-        if (x[i] != y[i])
-            return 0;
-    }
-    return 1;
-}
-
-FLINT_FORCE_INLINE void
-flint_mpn_negmod_n(mp_ptr res, mp_srcptr x, mp_srcptr m, mp_size_t n)
-{
-    if (flint_mpn_zero_p(x, n))
-        flint_mpn_zero(res, n);
-    else
-        mpn_sub_n(res, m, x, n);
-}
-
-FLINT_FORCE_INLINE void
-flint_mpn_addmod_n(mp_ptr res, mp_srcptr x, mp_srcptr y, mp_srcptr m, mp_size_t n)
-{
-    mp_limb_t cy;
-    cy = mpn_add_n(res, x, y, n);
-    if (cy || mpn_cmp(res, m, n) >= 0)
-        mpn_sub_n(res, res, m, n);
-}
-
-FLINT_FORCE_INLINE void
-flint_mpn_submod_n(mp_ptr res, mp_srcptr x, mp_srcptr y, mp_srcptr m, mp_size_t n)
-{
-    int cmp = (mpn_cmp(x, y, n) < 0);
-    mpn_sub_n(res, x, y, n);
-    if (cmp)
-        mpn_add_n(res, res, m, n);
-}
-
-/* assumes m <= n and y < m */
-FLINT_FORCE_INLINE void
-flint_mpn_addmod_n_m(mp_ptr res, mp_srcptr x, mp_srcptr y, mp_size_t yn, mp_srcptr m, mp_size_t n)
-{
-    mp_limb_t cy;
-    cy = mpn_add(res, x, n, y, yn);
-    if (cy || mpn_cmp(res, m, n) >= 0)
-        mpn_sub_n(res, res, m, n);
-}
-
-/* assumes m <= n and y < m */
-FLINT_FORCE_INLINE void
-flint_mpn_submod_n_m(mp_ptr res, mp_srcptr x, mp_srcptr y, mp_size_t yn, mp_srcptr m, mp_size_t n)
-{
-    int cmp = (flint_mpn_zero_p(x + yn, n - yn) && mpn_cmp(x, y, yn) < 0);
-    mpn_sub(res, x, n, y, yn);
-    if (cmp)
-        mpn_add_n(res, res, m, n);
-}
-
-FLINT_FORCE_INLINE void
-flint_mpn_negmod_2(mp_ptr res, mp_srcptr x, mp_srcptr m)
-{
-    if (x[0] == 0 && x[1] == 0)
-        res[1] = res[0] = 0;
-    else
-        sub_ddmmss(res[1], res[0], m[1], m[0], x[1], x[0]);
-}
-
-FLINT_FORCE_INLINE void
-flint_mpn_addmod_2(mp_ptr res, mp_srcptr x, mp_srcptr y, mp_srcptr m)
-{
-    mp_limb_t cy;
-    mp_limb_t m1 = m[1], m0 = m[0];
-    add_sssaaaaaa(cy, res[1], res[0], 0, x[1], x[0], 0, y[1], y[0]);
-    if (cy || (res[1] > m1 || (res[1] == m1 && res[0] >= m0)))
-        sub_ddmmss(res[1], res[0], res[1], res[0], m1, m0);
-}
-
-FLINT_FORCE_INLINE void
-flint_mpn_submod_2(mp_ptr res, mp_srcptr x, mp_srcptr y, mp_srcptr m)
-{
-    int cmp;
-    mp_limb_t m1 = m[1], m0 = m[0];
-    mp_limb_t x1 = x[1], x0 = x[0];
-    mp_limb_t y1 = y[1], y0 = y[0];
-    cmp = (x1 < y1) || (x1 == y1 && x0 < y0);
-    sub_ddmmss(res[1], res[0], x1, x0, y1, y0);
-    if (cmp)
-        add_ssaaaa(res[1], res[0], res[1], res[0], m1, m0);
-}
 
 void flint_mpn_mulmod_preinvn_2(mp_ptr r,
         mp_srcptr a, mp_srcptr b,
@@ -748,7 +623,6 @@ _gr_mpn_mod_sqr(mp_ptr res, mp_srcptr x, gr_ctx_t ctx)
     return _gr_mpn_mod_mul(res, x, x, ctx);
 }
 
-
 /* todo: check for 0? */
 int
 _gr_mpn_mod_mul_ui(mp_ptr res, mp_srcptr x, ulong y, gr_ctx_t ctx)
@@ -764,6 +638,8 @@ _gr_mpn_mod_mul_ui(mp_ptr res, mp_srcptr x, ulong y, gr_ctx_t ctx)
 }
 
 /* todo: check for 0? */
+#define UI_ABS_SI(x) (((slong)(x) < 0) ? (-(ulong)(x)) : ((ulong)(x)))
+
 int
 _gr_mpn_mod_mul_si(mp_ptr res, mp_srcptr x, slong y, gr_ctx_t ctx)
 {
