+++ src/mpn_mod/inlines.c
@@ -1,0 +1,14 @@
+/*
+    Copyright (C) 2015 William Hart
+
+    This file is part of FLINT.
+
+    FLINT is free software: you can redistribute it and/or modify it under
+    the terms of the GNU Lesser General Public License (LGPL) as published
+    by the Free Software Foundation; either version 3 of the License, or
+    (at your option) any later version.  See <https://www.gnu.org/licenses/>.
+*/
+
+#define MPN_MOD_INLINES_C
+
+#include "mpn_mod.h"
